+++ tests/integration/exclude.spec.js
@@ -1,5 +1,6 @@
 "use strict";
 
+const path       = require("path");
 const expect     = require("chai").expect;
 const simpleload = require("../../index.js");
 
@@ -7,47 +8,49 @@ describe("exclude", function () {
 
     it("should be possible to exclude given modules", function () {
 
-        var path = __dirname + "/../fixtures/dir_11",
-            loaded;
+        var modules = simpleload(path.join(__dirname, "/../fixtures/dir_11"), {
+            suffix: "service.js"
+        });
+        expect(modules.base).to.equal(require("../fixtures/dir_11/base.service"));
+        expect(modules.other).to.equal(require("../fixtures/dir_11/other.service"));
+        expect(Object.keys(modules)).to.have.length(2);
 
-        loaded = simpleload(path, { suffix: "service.js" });
-        expect(loaded.base).to.equal(require("../fixtures/dir_11/base.service"));
-        expect(loaded.other).to.equal(require("../fixtures/dir_11/other.service"));
-        expect(Object.keys(loaded)).to.have.length(2);
-
-
-        loaded = simpleload(path, { suffix: "service.js", exclude: "base" });
-        expect(loaded.other === require("../fixtures/dir_11/other.service"));
-        expect(Object.keys(loaded)).to.have.length(1);
+        modules = simpleload(path.join(__dirname, "/../fixtures/dir_11"), {
+            suffix: "service.js",
+            exclude: "base"
+        });
+        expect(modules.other === require("../fixtures/dir_11/other.service"));
+        expect(Object.keys(modules)).to.have.length(1);
 
     });
 
     it("should be possible to exclude given modules when array is passed", function () {
 
-        var path = __dirname + "/../fixtures/dir_11",
-            loaded;
+        var modules = simpleload(path.join(__dirname, "/../fixtures/dir_11"), {
+            suffix: "service.js"
+        });
+        expect(modules.base === require("../fixtures/dir_11/base.service"));
+        expect(modules.other === require("../fixtures/dir_11/other.service"));
+        expect(Object.keys(modules)).to.have.length(2);
 
-        loaded = simpleload(path, { suffix: "service.js" });
-        expect(loaded.base === require("../fixtures/dir_11/base.service"));
-        expect(loaded.other === require("../fixtures/dir_11/other.service"));
-        expect(Object.keys(loaded)).to.have.length(2);
-
-
-        loaded = simpleload(path, { suffix: "service.js", exclude: ["base"] });
-        expect(loaded.other === require("../fixtures/dir_11/other.service"));
-        expect(Object.keys(loaded)).to.have.length(1);
+        modules = simpleload(path.join(__dirname, "/../fixtures/dir_11"), {
+            suffix: "service.js",
+            exclude: ["base"]
+        });
+        expect(modules.other === require("../fixtures/dir_11/other.service"));
+        expect(Object.keys(modules)).to.have.length(1);
 
     });
 
     it("shouldn't exclude if falsy option was falled", function () {
 
-        var path = __dirname + "/../fixtures/dir_11",
-            loaded;
-
-        loaded = simpleload(path, { suffix: "service.js", exclude: false });
-        expect(loaded.base === require("../fixtures/dir_11/base.service"));
-        expect(loaded.other === require("../fixtures/dir_11/other.service"));
-        expect(Object.keys(loaded)).to.have.length(2);
+        var modules = simpleload(path.join(__dirname, "/../fixtures/dir_11"), {
+            suffix: "service.js",
+            exclude: false
+        });
+        expect(modules.base === require("../fixtures/dir_11/base.service"));
+        expect(modules.other === require("../fixtures/dir_11/other.service"));
+        expect(Object.keys(modules)).to.have.length(2);
 
     });