+++ tests/integration/global.spec.js
@@ -1,5 +1,6 @@
 "use strict";
 
+const path       = require("path");
 const expect     = require("chai").expect;
 const simpleload = require("../../index.js");
 
@@ -7,10 +8,10 @@ describe("global", function () {
 
     it("should be possible to expose the modules globally with the global flag set to true", function () {
 
-        var path = __dirname + "/../fixtures/dir_07",
-            modules;
-
-        modules = simpleload(path, { suffix: "model.js", global: true });
+        var modules = simpleload(path.join(__dirname, "/../fixtures/dir_07"), {
+            suffix: "model.js",
+            global: true
+        });
 
         expect(modules.token).to.exist;
         expect(modules.user).to.exist;
@@ -34,24 +35,17 @@ describe("global", function () {
 
     it("should be possible to override a global var if an override flag is passed", function () {
 
-        var path = __dirname + "/../fixtures/dir_07",
-            modules, exception;
+        global.token = "1234";
+        global.user = "bob";
 
-        modules = simpleload(path, { suffix: "model.js", global: true });
+        simpleload(path.join(__dirname, "/../fixtures/dir_07"), {
+            suffix: "model.js",
+            global: true,
+            override: true
+        });
 
-        expect(modules.token).to.exist;
-        expect(modules.user).to.exist;
-
-        expect(global.token).to.equal(modules.token);
-        expect(global.user).to.equal(modules.user);
-
-        try {
-            simpleload(path, { suffix: "model.js", global: true, override: true });
-        } catch (e) {
-            exception = e;
-        }
-
-        expect(exception).to.not.exist;
+        expect(global.token).to.not.equal("1234");
+        expect(global.user).to.not.equal("bob");
 
         delete global.token;
         delete global.user;
@@ -63,31 +57,41 @@ describe("global", function () {
 
     it("should throw an error if a global var is going to be overriden but no flag is passed", function () {
 
-        var path = __dirname + "/../fixtures/dir_07",
-            modules, exception;
+        expect(global.Model).not.to.exist;
 
-        modules = simpleload(path, { suffix: "model.js", namespace: "models", global: true });
+        var modules = simpleload(path.join(__dirname, "/../fixtures/dir_07"), {
+            suffix: "model.js",
+            namespace: "Model",
+            global: true
+        });
 
         expect(modules.token).to.exist;
         expect(modules.user).to.exist;
 
-        expect(global.models.token).to.equal(modules.token);
-        expect(global.models.user).to.equal(modules.user);
+        expect(global.Model.token).to.equal(modules.token);
+        expect(global.Model.user).to.equal(modules.user);
 
         expect(function () {
-            simpleload(path, { suffix: "model.js", namespace: "models", global: true });
+            simpleload(path.join(__dirname, "/../fixtures/dir_07"), {
+                suffix: "model.js",
+                namespace: "Model",
+                global: true
+            });
         }).to.throw();
 
-        delete global.models;
+        delete global.Model;
 
     });
 
     it("should be possible to add modules to chosen global namespace", function () {
 
-        var path = __dirname + "/../fixtures/dir_07",
-            modules;
+        expect(global.Model).not.to.exist;
 
-        modules = simpleload(path, { suffix: "model.js", global: true, namespace: "Model" });
+        var modules = simpleload(path.join(__dirname, "/../fixtures/dir_07"), {
+            suffix: "model.js",
+            global: true,
+            namespace: "Model"
+        });
 
         expect(modules.token).to.exist;
         expect(modules.user).to.exist;
@@ -98,8 +102,6 @@ describe("global", function () {
         delete global.Model.token;
         delete global.Model.user;
         delete global.Model;
-
-        expect(global.Model).to.not.exist;
 
     });