+++ HISTORY.md
@@ -2,13 +2,14 @@
 ==================
 
   * add: load more formats via the `extension` option
+  * add: more `recursive` flag specs
 
 0.6.2 / 2016-01-25
 ==================
 
   * add: coverage via istanbul
   * add: npm run coverage in package.json
-  * add: decorator specs
+  * add: `decorator` specs
 
 0.6.1 / 2015-11-15
 ==================